+++ Dictation/natural-mode.talon
@@ -1,0 +1,4 @@
+mode: dictation
+-
+
+<user.raw_prose>:           user.gpt_dictation_insert(raw_prose)
