--- NaturalDictation/mode-switch.talon
@@ -1,6 +1,0 @@
-natural mode:
-    mode.disable("sleep")
-    mode.disable("command")
-    mode.disable("dictation")
-    user.code_clear_language_mode()
-    user.gdb_disable()
