--- NaturalDictation/natural-mode.talon
@@ -1,4 +1,0 @@
-mode: NaturalSpeech
--
-
-<user.text>:                user.gpt_apply_prompt()
